requirements.txt
pygame==2.5.2
numpy==1.26.4
scikit-learn==1.4.1
tensorflow==2.16.0
Pillow==10.1.0
pywebview==5.0.2
